IMPROVED determination of ground position under character
The ground position calculation now takes composite placeables into account and returns the top of the part the character actually stands on. This allows for much smoother height transitions when walking up/down stairs.

Additional code could be added to account for sloped parts, where the top is not the same height at every position.