ADDED scaling support to the screen class.
The purpose is to keep using the desktop resolution and aspect ratio
when running in fullscreen mode, while also keeping a fairly equal
sized view of the game world. For that purpose, a minimum and maximum
view size has been defined, so that the actual view size can be
magnified to the desktop resolution by an integer scaling factor.
x# Please enter the commit message for your changes. Lines starting