Fixed lack of return in non-void function
The is...below() function appears to me to be complete, and so
I did not presume to return true or false in the final case.
Instead I added an exception.  This may impact performance slightly,
and may need to be removed, and given a real true/false, if
all tests above fail.