cmake: added override setting for compiling swig wrapper code
SWIG wrapper code has so many useless warnings in it, that it causes
us to turn off -Werror during compiles, just to get a finished product.

Another product's bugs shouldn't encourage bugs in our own code.

This commit adds the flag: ADDITIONAL_SWIG_CXXFLAGS, which can be
set on the cmake command line, which will be added to all python
code.  This catches the py-runtime directory, and the py-wrappers
directory.  Unfortunately, it also catches py-wrappers/py_main.cc, which
is our code as well, but that is only one file.

To enable -Wall -Werror while skipping SWIG errors, use:

cmake \
	-DCMAKE_C_FLAGS:STRING="-Wall -Werror" \
	-DCMAKE_CXX_FLAGS:STRING="-Wall -Werror" \
	-DADDITIONAL_SWIG_CXXFLAGS:STRING="-Wno-error" \
	..

The -Wno-error is added after the usual flags, allowing the compile to
succeed for the SWIG wrapper code.