world: removed the exception from renderer.cc
After analyzing the code, all possible cases are handled in the switch
statement, but the compiler is not always smart enough to figure that
out (and it took me a while too).  :-)  So, for the "should never happen"
case, add a default: case under case 0, which displays debug output,
which is a suitable response.

This removes the exception I added before to fix compile issues, and
adonthell should now compile again when exception are turned off.