Fixed header include priorities, to avoid confusion of system headers
This commit avoids confusion between adonthell headers and system headers
with the same name, and sometimes even the same directory.

The key is knowing that whenever #include is used with quotes, such as
as the file doing the including.  Repeat, in the same directory as the
file _doing_ the including.

So, for example, if an application does: #include <adonthell/blah/blah.h>
and that blah.h header does: #include "help.h"
then the compiler will search for help.h in adonthell/blah/ *first*,
because that is where blah.h lives.

This is not the case for #include <> style includes, which search the
path list first.

Therefore, we end up with these following simple rules:

	1) if the header you want is in the same directory as the file
	   you are working with, use quotes

	2) if the header you want is always relative to the directory of
	   the file you are working with, use quotes

	3) for everything else, use angle brackets and the full path of
	   the header, including the library name

So, for example, if you're working on src/audio/sound.cc and need to
include sound.h, use: #include "sound.h", with no path, because sound.h
is in the same directory as sound.cc.

If you are working on src/audio/audio.cc and need to include the file
src/base/base.h, use: #include <adonthell/base/base.h>, since the file
is outside of the module you wish to use.

When Adonthell is installed as a library, its headers will also now be
installed under /usr/include/adonthell-<version>/adonthell/...

Therefore, when building an application that depends on, say, adonthell-1.1,
pass -I/usr/include/adonthell-1.1 to the application's compiler, which
will make its #include <adonthell/base/base.h> work properly.