FIXED issues with mutating tuple surfaced in eventtest.py.
The event test script showed lots of errors similar to the following:

Traceback (most recent call last):
  File "data/game_events/time_events.py", line 41, in callback_2
    print "#2", self.counter, evt.time (), base.Timer.uptime (),
lst.get_event ().repeat ()
SystemError: Objects/tupleobject.c:142: bad argument to internal
function

Appearently, python tuples are supposed to be immutable, so it is a bad
idea to use the same tuple and just exchange every parameter. We still
update the event parameter every time the callback is executed, but this
does not trigger an error in the test script.